Cap the number of physics ticks per frame.
The old approach was to spam as much physics as necessary in order to
keep up, but this had cascading failure if physics ticks ended up taking
longer than the time between frames and the whole game would freeze.

Now, the number of ticks per frame is capped at 6 (150% speed). This
allows a small amount of lag to be caught up from, but degrades to
simulating in slow motion if the processing can't keep up.